Extract elm fixture
